requirements.txt
pyTelegramBotApi
requests
python-decouple-3.3
